+++ package.json
@@ -11,8 +11,8 @@
     "scripts": {
         "build": "node_modules\\.bin\\tsc",
         "test": "mocha --reporter mocha-multi-reporters --reporter-options configFile=src/test/mochaConfig.json -r ts-node/register src/test/libraryTest/**/*.ts",
-        "test:benchmark": "npx ts-node src\\test\\resourceTest\\benchmark\\createBenchmarks.ts",
         "test:resources": "mocha --reporter mocha-multi-reporters --reporter-options configFile=src/test/mochaConfig.json -r ts-node/register src/test/resourceTest/**/*.ts",
+        "script:benchmark": "npx ts-node src\\scripts\\createBenchmarks.ts",
         "lint": "eslint src --ext ts",
         "prepublishOnly": "git clean -xdf && npm install-clean && npm run lint && npm run build && npm run test && npm run test:resources"
     },
